// extension/pplcz/src/Admin/MuiAdmin/src/schema.ts
export interface components {
  schemas: {
    ProductModel: {
      pplConfirmAge15?: boolean | null;
      pplConfirmAge18?: boolean | null;
      pplDisabledParcelBox?: boolean | null;
      pplDisabledAlzaBox?: boolean;
      pplDisabledParcelShop?: boolean | null;
      pplDisabledTransport?: string[] | null;
    };
    CategoryModel: {
      pplDisabledParcelBox?: boolean | null;
      pplDisabledAlzaBox?: boolean;
      pplDisabledParcelShop?: boolean | null;
      pplDisabledTransport?: string[] | null;
      pplConfirmAge15?: boolean | null;
      pplConfirmAge18?: boolean | null;
    };
// ...
    ShipmentMethodModel: {
      code: string;
      title: string;
      codAvailable: boolean;
      parcelRequired: boolean;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}